requirements.txt
asgiref==3.3.*
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
Django==3.2.*
django-common-helpers==0.9.2
django-crispy-forms==1.9.*
gunicorn==20.1.0
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
pytz==2020.1
psycopg2-binary
whitenoise==5.2.0
requests==2.26.0
six==1.15.0
sqlparse==0.4.*
toml==0.10.1
urllib3==1.26.6
wrapt==1.12.1
django-tinymce==3.4.*
